+++ src/TokenizerFwf.h
@@ -33,6 +33,7 @@ private:
   Token fieldToken(SourceIterator begin, SourceIterator end, bool hasNull);
 
   bool isComment(const char* cur) const;
+  bool isEmpty() const;
 };
 
 #endif