+++ tests/testthat/test-read-fwf.R
@@ -204,3 +204,9 @@ test_that("read_table can read from a pipe (552)", {
   x <- read_table(pipe("echo a b c && echo 1 2 3 && echo 4 5 6"), progress = FALSE)
   expect_equal(x$a, c(1, 4))
 })
+
+test_that("read_table does not duplicate header rows for leading whitespace (747)", {
+  x <- read_table("\nfoo bar\n1   2")
+  expect_equal(nrow(x), 1)
+  expect_equal(x$foo, 1)
+})