+++ tests/testthat/test-read-table.R
@@ -43,3 +43,9 @@ test_that("read_table2 can read from a pipe (552)", {
   x <- read_table2(pipe("echo a b c && echo 1 2 3 && echo 4 5 6"))
   expect_equal(x$a, c(1, 4))
 })
+
+test_that("read_table2 does not duplicate header rows for leading whitespace (747)", {
+  x <- read_table2("\nfoo bar\n1   2")
+  expect_equal(nrow(x), 1)
+  expect_equal(x$foo, 1)
+})